Fixed button
The Clear parent button was missing and set up somewhat silly before.
thanks to @novice-coding-butterfly for pointing it out about 2 weeks ago.

I did notice it was gone but I didn't go back to fix it myself heh.
