+++ manifest.json
@@ -1,7 +1,7 @@
 {
   "name": "Mektools",
   "author": "Meku Maki, Shino Mythmaker",
-"version": [1, 2, 4],
+"version": [1, 2, 5],
 "feature_name": "dev",
 "feature_patch": 0,
   "blender": {
